refactor: move integration of Segment to integral.py file
