+++ tests/geometry/test_segment.py
@@ -2,13 +2,11 @@
 This file contains tests functions to test the module polygon.py
 """
 
-import math
 from fractions import Fraction
 
-import numpy as np
 import pytest
 
-from shapepy.geometry.segment import IntegratePlanar, Math, Segment
+from shapepy.geometry.segment import Segment
 
 
 @pytest.mark.order(13)
@@ -56,114 +54,6 @@ class TestDerivate:
         depends=[
             "TestDerivate::test_begin",
             "TestDerivate::test_planar_bezier",
-        ]
-    )
-    def test_end(self):
-        pass
-
-
-class TestIntegrate:
-    @pytest.mark.order(13)
-    @pytest.mark.dependency(depends=["test_begin"])
-    def test_begin(self):
-        pass
-
-    @pytest.mark.order(13)
-    @pytest.mark.timeout(10)
-    @pytest.mark.dependency(depends=["TestIntegrate::test_begin"])
-    def test_lenght(self):
-        points = [(0, 0), (1, 0)]
-        curve = Segment(points)
-        assert IntegratePlanar.lenght(curve) == 1
-
-        points = [(1, 0), (0, 0)]
-        curve = Segment(points)
-        assert IntegratePlanar.lenght(curve) == 1
-
-        points = [(0, 1), (1, 0)]
-        curve = Segment(points)
-        assert abs(IntegratePlanar.lenght(curve) - np.sqrt(2)) < 1e-9
-
-    @pytest.mark.order(13)
-    @pytest.mark.timeout(10)
-    @pytest.mark.dependency(depends=["TestIntegrate::test_begin"])
-    def test_winding_triangles(self):
-        curve = Segment([(1, 0), (2, 0)])
-        wind = IntegratePlanar.winding_number(curve)
-        assert wind == 0
-
-        curve = Segment([(1, 0), (1, 1)])
-        wind = IntegratePlanar.winding_number(curve)
-        assert abs(wind - 0.125) < 1e-9
-
-        curve = Segment([(1, 0), (0, 1)])
-        wind = IntegratePlanar.winding_number(curve)
-        assert abs(wind - 0.25) < 1e-9
-
-        curve = Segment([(1, 0), (-0.5, np.sqrt(3) / 2)])
-        wind = IntegratePlanar.winding_number(curve)
-        assert abs(3 * wind - 1) < 1e-9
-
-    @pytest.mark.order(13)
-    @pytest.mark.timeout(10)
-    @pytest.mark.dependency(
-        depends=[
-            "TestIntegrate::test_begin",
-            "TestIntegrate::test_winding_triangles",
-        ]
-    )
-    def test_winding_unit_circle(self):
-        ntests = 1000
-        maxim = 0
-        for k in range(ntests):
-            angle0 = np.random.uniform(0, 2 * np.pi)
-            good_wind = np.random.uniform(-0.5, 0.5)
-            angle1 = angle0 + 2 * np.pi * good_wind
-            point0 = (np.cos(angle0), np.sin(angle0))
-            point1 = (np.cos(angle1), np.sin(angle1))
-            curve = Segment([point0, point1])
-            test_wind = IntegratePlanar.winding_number(curve)
-            diff = abs(good_wind - test_wind)
-            maxim = max(maxim, diff)
-            assert abs(good_wind - test_wind) < 1e-9
-
-    @pytest.mark.order(13)
-    @pytest.mark.timeout(10)
-    @pytest.mark.dependency(
-        depends=[
-            "TestIntegrate::test_begin",
-            "TestIntegrate::test_winding_triangles",
-            "TestIntegrate::test_winding_unit_circle",
-        ]
-    )
-    def test_winding_regular_polygon(self):
-        for nsides in range(3, 10):
-            angles = np.linspace(0, math.tau, nsides + 1)
-            ctrlpoints = np.vstack([np.cos(angles), np.sin(angles)]).T
-            pairs = zip(ctrlpoints[:-1], ctrlpoints[1:])
-            curves = tuple(Segment(pair) for pair in pairs)
-            for curve in curves:
-                wind = IntegratePlanar.winding_number(curve)
-                assert abs(nsides * wind - 1) < 1e-2
-
-        for nsides in range(3, 10):
-            angles = np.linspace(math.tau, 0, nsides + 1)
-            ctrlpoints = np.vstack([np.cos(angles), np.sin(angles)]).T
-            pairs = zip(ctrlpoints[:-1], ctrlpoints[1:])
-            curves = tuple(Segment(pair) for pair in pairs)
-            for curve in curves:
-                wind = IntegratePlanar.winding_number(curve)
-                assert abs(nsides * wind + 1) < 1e-2
-
-    @pytest.mark.order(13)
-    @pytest.mark.timeout(10)
-    @pytest.mark.dependency(
-        depends=[
-            "TestIntegrate::test_begin",
-            "TestIntegrate::test_lenght",
-            "TestIntegrate::test_winding_triangles",
-            "TestIntegrate::test_winding_unit_circle",
-            "TestIntegrate::test_winding_regular_polygon",
         ]
     )
     def test_end(self):
@@ -334,7 +224,6 @@ def test_print():
         "test_begin",
         "test_build",
         "TestDerivate::test_end",
-        "TestIntegrate::test_end",
         "TestOperations::test_end",
         "TestContains::test_end",
         "TestSplitUnite::test_end",
